+++ e2e/allowance.e2e.ts
@@ -1,14 +1,14 @@
+import { ChainId, CoinKey, findDefaultToken, Token } from '@lifi/types'
+import BigNumber from 'bignumber.js'
+import { constants, providers, Signer, Wallet } from 'ethers'
 import {
   approveToken,
   bulkGetTokenApproval,
   getTokenApproval,
   revokeTokenApproval,
 } from '../src/allowance'
-import { constants, providers, Signer, Wallet } from 'ethers'
-import { ChainId, CoinKey, findDefaultToken, Token } from '@lifinance/types'
-import BigNumber from 'bignumber.js'
-import { sleep } from '../src/utils/utils'
 import { getApproved, setApproval } from '../src/allowance/utils'
+import { sleep } from '../src/utils/utils'
 
 const USDC_ON_DAI = findDefaultToken(CoinKey.USDC, ChainId.DAI)
 const USDT_ON_DAI = findDefaultToken(CoinKey.USDT, ChainId.DAI)
@@ -66,13 +66,13 @@ describe('allowance e2e tests', () => {
 
       console.log('setting new approval')
 
-      await approveToken(
+      await approveToken({
         signer,
-        USDC_ON_DAI,
-        SUSHISWAP_APPROVAL_ADDRESS_ON_DAI,
+        token: USDC_ON_DAI,
+        approvalAddress: SUSHISWAP_APPROVAL_ADDRESS_ON_DAI,
         amount,
-        false
-      )
+        infiniteApproval: false,
+      })
 
       await checkSetApproval(USDC_ON_DAI.address, amount)
     })
@@ -82,13 +82,13 @@ describe('allowance e2e tests', () => {
 
       console.log('setting new approval')
 
-      await approveToken(
+      await approveToken({
         signer,
-        USDC_ON_DAI,
-        SUSHISWAP_APPROVAL_ADDRESS_ON_DAI,
+        token: USDC_ON_DAI,
+        approvalAddress: SUSHISWAP_APPROVAL_ADDRESS_ON_DAI,
         amount,
-        true
-      )
+        infiniteApproval: true,
+      })
 
       await checkSetApproval(
         USDC_ON_DAI.address,
@@ -103,13 +103,13 @@ describe('allowance e2e tests', () => {
 
       console.log('setting new approval')
 
-      await approveToken(
+      await approveToken({
         signer,
-        USDC_ON_DAI,
-        SUSHISWAP_APPROVAL_ADDRESS_ON_DAI,
+        token: USDC_ON_DAI,
+        approvalAddress: SUSHISWAP_APPROVAL_ADDRESS_ON_DAI,
         amount,
-        false
-      )
+        infiniteApproval: false,
+      })
 
       await sleep(2000)
 
@@ -117,11 +117,11 @@ describe('allowance e2e tests', () => {
 
       console.log('revoking approval')
 
-      await revokeTokenApproval(
+      await revokeTokenApproval({
         signer,
-        USDC_ON_DAI,
-        SUSHISWAP_APPROVAL_ADDRESS_ON_DAI
-      )
+        token: USDC_ON_DAI,
+        approvalAddress: SUSHISWAP_APPROVAL_ADDRESS_ON_DAI,
+      })
 
       await sleep(2000)
 
@@ -135,13 +135,13 @@ describe('allowance e2e tests', () => {
 
       console.log('setting new approval')
 
-      await approveToken(
+      await approveToken({
         signer,
-        USDC_ON_DAI,
-        SUSHISWAP_APPROVAL_ADDRESS_ON_DAI,
+        token: USDC_ON_DAI,
+        approvalAddress: SUSHISWAP_APPROVAL_ADDRESS_ON_DAI,
         amount,
-        false
-      )
+        infiniteApproval: false,
+      })
 
       await sleep(2000)
 
@@ -168,13 +168,13 @@ describe('allowance e2e tests', () => {
 
       console.log('setting new approval')
 
-      await approveToken(
+      await approveToken({
         signer,
-        USDT_ON_DAI,
-        SUSHISWAP_APPROVAL_ADDRESS_ON_DAI,
+        token: USDT_ON_DAI,
+        approvalAddress: SUSHISWAP_APPROVAL_ADDRESS_ON_DAI,
         amount,
-        false
-      )
+        infiniteApproval: false,
+      })
 
       await sleep(2000)