+++ src/Lifi.ts
@@ -19,7 +19,7 @@ import {
   TokensResponse,
   ToolsRequest,
   ToolsResponse,
-} from '@lifinance/types'
+} from '@lifi/types'
 import { Signer } from 'ethers'
 import {
   approveToken,
@@ -278,6 +278,15 @@ export default class LIFI {
         return clonedRoute
       }
     }
+    // remove last (failed) process
+    for (let index = 0; index < clonedRoute.steps.length; index++) {
+      const stepHasFailed =
+        clonedRoute.steps[index].execution?.status === 'FAILED'
+
+      if (stepHasFailed) {
+        clonedRoute.steps[index].execution?.process.pop()
+      }
+    }
 
     return this.executeSteps(signer, clonedRoute, settings)
   }
@@ -404,7 +413,11 @@ export default class LIFI {
     }
 
     if (!isToken(token)) {
-      throw new ValidationError('Invalid token passed.')
+      throw new ValidationError(
+        `Invalid token passed: address "${
+          (token as Token).address
+        }" on chainId "${(token as Token).chainId}"`
+      )
     }
 
     return balances.getTokenBalance(walletAddress, token)
@@ -425,8 +438,11 @@ export default class LIFI {
       throw new ValidationError('Missing walletAddress.')
     }
 
-    if (tokens.filter((token) => !isToken(token)).length) {
-      throw new ValidationError('Invalid token passed.')
+    const invalidTokens = tokens.filter((token) => !isToken(token))
+    if (invalidTokens.length) {
+      throw new ValidationError(
+        `Invalid token passed: address "${invalidTokens[0].address}" on chainId "${invalidTokens[0].chainId}"`
+      )
     }
 
     return balances.getTokenBalances(walletAddress, tokens)
@@ -448,8 +464,11 @@ export default class LIFI {
     }
 
     const tokenList = Object.values(tokensByChain).flat()
-    if (tokenList.filter((token) => !isToken(token)).length) {
-      throw new ValidationError('Invalid token passed.')
+    const invalidTokens = tokenList.filter((token) => !isToken(token))
+    if (invalidTokens.length) {
+      throw new ValidationError(
+        `Invalid token passed: address "${invalidTokens[0].address}" on chainId "${invalidTokens[0].chainId}"`
+      )
     }
 
     return balances.getTokenBalancesForChains(walletAddress, tokensByChain)