+++ src/execution/utils.ts
@@ -1,34 +1,35 @@
 import {
-  BridgeTool,
-  ExchangeTools,
   ProcessType,
   Status,
+  StatusMessage,
   StatusResponse,
   Step,
+  Substatus,
 } from '@lifinance/types'
 import BigNumber from 'bignumber.js'
 import { ChainId } from '..'
+import { StatusManager } from '..'
+
 import ApiService from '../services/ApiService'
-import { LifiErrorCode, ServerError, TransactionError } from '../utils/errors'
-import { getSlippageNotMetMessage } from '../utils/parseError'
+import { ServerError } from '../utils/errors'
 import { repeatUntilDone } from '../utils/utils'
 
 const TRANSACTION_HASH_OBSERVERS: { [txHash: string]: Promise<any> } = {}
 
 export async function waitForReceivingTransaction(
-  tool: BridgeTool | ExchangeTools,
-  fromChainId: ChainId,
-  toChainId: ChainId,
-  txHash: string
+  txHash: string,
+  statusManager: StatusManager,
+  processType: ProcessType,
+  step: Step
 ): Promise<StatusResponse> {
   const getStatus = (): Promise<StatusResponse | undefined> =>
     new Promise(async (resolve, reject) => {
       let statusResponse: StatusResponse
       try {
         statusResponse = await ApiService.getStatus({
-          bridge: tool,
-          fromChain: fromChainId,
-          toChain: toChainId,
+          bridge: step.tool,
+          fromChain: step.action.fromChainId,
+          toChain: step.action.toChainId,
           txHash,
         })
       } catch (e: any) {
@@ -40,6 +41,16 @@ export async function waitForReceivingTransaction(
         case 'DONE':
           return resolve(statusResponse)
         case 'PENDING':
+          statusManager?.updateProcess(step, processType, 'PENDING', {
+            substatus: statusResponse.substatus,
+            substatusMessage:
+              statusResponse.substatusMessage ||
+              getSubstatusMessage(
+                statusResponse.status,
+                statusResponse.substatus
+              ),
+          })
+          return resolve(undefined)
         case 'NOT_FOUND':
           return resolve(undefined)
         case 'FAILED':
@@ -67,8 +78,8 @@ export async function waitForReceivingTransaction(
 const processMessages: Record<ProcessType, Partial<Record<Status, string>>> = {
   TOKEN_ALLOWANCE: {
     STARTED: 'Setting token allowance.',
-    PENDING: 'Waiting for token allowance approval.',
-    DONE: 'Token allowance approved.',
+    PENDING: 'Waiting for token allowance.',
+    DONE: 'Token allowance set.',
   },
   SWITCH_CHAIN: {
     PENDING: 'Chain switch required.',
@@ -91,6 +102,32 @@ const processMessages: Record<ProcessType, Partial<Record<Status, string>>> = {
     DONE: 'Funds received.',
   },
   TRANSACTION: {},
+}
+const substatusMessages: Record<
+  StatusMessage,
+  Partial<Record<Substatus, string>>
+> = {
+  PENDING: {
+    BRIDGE_NOT_AVAILABLE: 'Bridge communication is temporarily unavailable.',
+    CHAIN_NOT_AVAILABLE: 'RPC communication is temporarily unavailable.',
+    NOT_PROCESSABLE_REFUND_NEEDED:
+      'The transfer cannot be completed successfully. A refund operation is required.',
+    UNKNOWN_ERROR:
+      'An unexpected error occurred. Please seek assistance in the LI.FI discord server.',
+    WAIT_SOURCE_CONFIRMATIONS:
+      'The bridge deposit has been received. The bridge is waiting for more confirmations to start the off-chain logic.',
+    WAIT_DESTINATION_TRANSACTION:
+      'The bridge off-chain logic is being executed. Wait for the transaction to appear on the destination chain.',
+  },
+  DONE: {
+    PARTIAL:
+      'Some of the received tokens are not the requested destination tokens.',
+    REFUNDED: 'The tokens were refunded to the sender address.',
+    COMPLETED: 'The transfer is complete.',
+  },
+  FAILED: {},
+  INVALID: {},
+  NOT_FOUND: {},
 }
 
 export function getProcessMessage(
@@ -119,4 +156,14 @@ export function updatedStepMeetsSlippageConditions(
   } else {
     return false
   }
+}
+export function getSubstatusMessage(
+  status: StatusMessage,
+  substatus?: Substatus
+): string | undefined {
+  if (!substatus) {
+    return
+  }
+  const message = substatusMessages[status][substatus]
+  return message
 }